requirements.txt
alabaster==0.7.12
amqp==5.0.2
appdirs==1.4.3
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
billiard==3.6.3.0
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.3.1
cssselect==1.1.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.4
Faker==5.6.1
django-celery-beat==2.1.0
django-cors-headers==3.6.0
django-extensions==3.1.0
django-filter==2.4.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
imagesize==1.2.0
incremental==17.5.0
iniconfig==1.1.1
ipaddr==2.2.0
ipython==7.19.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
lockfile==0.12.2
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
parsel==1.6.0
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
progress==1.5
prompt-toolkit==3.0.8
Protego==0.1.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.3
PyHamcrest==2.0.2
PyJWT==2.0.0
pyOpenSSL==20.0.0
pyparsing==2.4.6
pytest==6.2.1
pytest-django==4.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
queuelib==1.5.0
repackage==0.7.3
requests==2.22.0
retrying==1.3.3
Scrapy==2.4.1
service-identity==18.1.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
Twisted==20.3.0
urllib3==1.25.8
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
zope.interface==5.2.0
